chore: use only one environment file
In a real world app you shouldn't build seperate artifacts
for each environment/ You should build one artifact, test it
and then deploy the same artifact to all environments.
This is why the msal config is externalized in this application.
It can be updated and validated seperately from the app.